Move signing related test helpers to own module
